#4 first pass to refactor classes to match python api * used folders in the project to group classes similar to the python file sets
