+++ OMF/OMF.csproj
@@ -58,29 +58,35 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Objects\ClassBase.cs" />
+    <Compile Include="base\ClassBase.cs" />
+    <Compile Include="base\ProjectElementData.cs" />
+    <Compile Include="base\ProjectElementGeometry.cs" />
+    <Compile Include="base\ProjectElement.cs" />
+    <Compile Include="base\Project.cs" />
+    <Compile Include="base\Version.cs" />
+    <Compile Include="Objects\InvalidOmfException.cs" />
     <Compile Include="Objects\Helpers.cs" />
     <Compile Include="Objects\AttributeDefinitionArray.cs" />
     <Compile Include="Objects\AttributeArray.cs" />
-    <Compile Include="Objects\VolumeGridGeometry.cs" />
-    <Compile Include="OMFWriter.cs" />
-    <Compile Include="OMFReader.cs" />
-    <Compile Include="Objects\DateAndDescriptionBase.cs" />
-    <Compile Include="Objects\DateBase.cs" />
+    <Compile Include="volume\VolumeGridGeometry.cs" />
+    <Compile Include="file\OMFWriter.cs" />
+    <Compile Include="file\OMFReader.cs" />
+    <Compile Include="base\ContentModel.cs" />
+    <Compile Include="base\UidModel.cs" />
     <Compile Include="Objects\Legend.cs" />
     <Compile Include="Objects\ObjectFactory.cs" />
-    <Compile Include="Objects\IObject.cs" />
+    <Compile Include="base\IObject.cs" />
     <Compile Include="Objects\Int3Array.cs" />
-    <Compile Include="Objects\LineSetGeometry.cs" />
+    <Compile Include="lineset\LineSetGeometry.cs" />
     <Compile Include="Objects\ScalarData.cs" />
     <Compile Include="Objects\ScalarArray.cs" />
     <Compile Include="Objects\ColorArray.cs" />
     <Compile Include="Objects\StringArray.cs" />
-    <Compile Include="Objects\LineSetElement.cs" />
+    <Compile Include="lineset\LineSetElement.cs" />
     <Compile Include="Objects\Int2Array.cs" />
-    <Compile Include="Objects\VolumeElement.cs" />
-    <Compile Include="Objects\SurfaceElement.cs" />
-    <Compile Include="Objects\SurfaceGeometry.cs" />
+    <Compile Include="volume\VolumeElement.cs" />
+    <Compile Include="surface\SurfaceElement.cs" />
+    <Compile Include="surface\SurfaceGeometry.cs" />
     <Compile Include="Objects\Vector3Array.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="Properties\Resources.resx">
@@ -105,12 +111,13 @@
       <DependentUpon>Settings.settings</DependentUpon>
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
     </Compile>
-    <Compile Include="Objects\PointSetElement.cs" />
-    <Compile Include="Objects\PointSetGeometry.cs" />
+    <Compile Include="pointset\PointSetElement.cs" />
+    <Compile Include="pointset\PointSetGeometry.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
        Other similar extension points exist, see Microsoft.Common.targets.
