+++ OMF/Objects/ColorArray.cs
@@ -5,9 +5,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
-namespace OMF.Objects
+namespace OMF.Base
 {
-    public class ColorArray : DateBase, IObject
+    public class ColorArray : UidModel, IObject
     {
         public string[] array { get; set; }
