+++ OMF/Objects/ScalarData.cs
@@ -6,9 +6,9 @@ using System.Threading.Tasks;
 using Newtonsoft.Json;
 using System.IO;
 
-namespace OMF.Objects
+namespace OMF.Base
 {
-    public class ScalarData : DateBase, IObject
+    public class ScalarData : UidModel, IObject
     {
         public string array { get; set; }
         public string location { get; set; }
