+++ OMF/Objects/StringArray.cs
@@ -5,9 +5,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
-namespace OMF.Objects
+namespace OMF.Base
 {
-    public class StringArray : DateBase, IObject
+    public class StringArray : UidModel, IObject
     {
         public string[] array { get; set; }
