+++ OMF/base/ClassBase.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OMF.Base
+{
+    public class ClassBase
+    {
+        public ClassBase()
+        {
+            __class__ = this.GetType().Name;
+        }
+        public string __class__ { get; set; }
+    }
+}
