+++ OMF/base/ProjectElementGeometry.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OMF
+{
+    public class ProjectElementGeometry: UidModel
+    {
+        public Tuple<double,double,double> origin { get; set; }
+
+    }
+}
